Ref/Deref evaluation
Adds Address pterm branch to identify addresses, but this is strange because we don't want to actually be able to construct such terms.

APS project handles this by having alloc in the AST, which then returns an InA. So, if I create a lexer/parser, I could take a similar approach.